recommend for user/requirements.txt
python-dotenv
openai
groq
sentence-transformers
qdrant-client
pillow
